requirements.txt
paho-mqtt
datetime
urllib3
speedtest-cli